src/main/java/edu/tdd/service/UserService.java
package edu.tdd.service;

import edu.tdd.dao.UserDao;
import edu.tdd.dto.CreateUserDto;
import edu.tdd.dto.UserDto;
import edu.tdd.exception.ValidationException;
import edu.tdd.mapper.CreateUserMapper;
import edu.tdd.mapper.UserMapper;
import edu.tdd.validator.CreateUserValidator;
import lombok.RequiredArgsConstructor;
import lombok.SneakyThrows;

import java.util.Optional;

@RequiredArgsConstructor
public class UserService {

    private final CreateUserValidator createUserValidator;
    private final UserDao userDao;
    private final CreateUserMapper createUserMapper;
    private final UserMapper userMapper;

    public Optional<UserDto> login(String email, String password) {
        return userDao.findByEmailAndPassword(email, password)
                .map(userMapper::map);
    }

    @SneakyThrows
    public UserDto create(CreateUserDto userDto) {
        var validationResult = createUserValidator.validate(userDto);
        if (validationResult.hasErrors()) {
            throw new ValidationException(validationResult.getErrors());
        }
        var userEntity = createUserMapper.map(userDto);
        userDao.save(userEntity);

        return userMapper.map(userEntity);
    }
}
